Translation checked -La Traditora
Many changes to do with specifying the expansion of abbreviations.
A few words bracketed, to indicate the lack of correspondence to the literal words of the text.
A few words mended because of inopportune word breaks.
Some footnotes suggested (in comments for now), to denote words possibly crossed out in the text.