Translation checked -Baramattio
Added precise markings for abbreviation expansions.
Was able to fill in several doubtful or unclear characters.
Changed a few words; changed some word positions.
In one spot (about line 173), changed the word "from" to "to", based on word choice in the original; made a comment in the text, and will enter as an issue.